Simulation of a real-world Banking software
This project showcases a simulation of a real-world banking software. A team of 3 students and I  came up with the Idea of simulating the workings of a real word bank software. We created a hypothetical bank and called it Revolve Bank . The bank had a hierarchical leadership structure with the Bank manager at the top  all the way down to a bank teller and customers. This hierarchical  system meant different access and management levels to the customer accounts. We also included different functionalities to a customer’s account such as money withdrawal, depositing, transferring and loan requests. At this point, I am responsible for doing all the improvements and Implementation of new features. All commits included,  will bear specific detailed comments about the new features and improvements that I am working on.