Covid-19 Spread Simulation
In this project, I sought to create a program to simulate the concept behind the spread of COVID 19. In the program I created attributes of  everyday citizens: some were healthy, some had  underlying diseases,  others wore masks, others did not, and some were healthcare workers. These attributes affected their chances of either catching, or spreading the disease once they already have it.  Quite superficially, we would expect a senior, wearing no mask  and having underlying conditions to be more susceptible than a healthy , young person who puts on masks in public places. With this logic, I created a virtual world where we could see in real-time, the  hotspots of the virus and how the it continued to either diminish or grow, and at what rates.